requirements.txt
prometheus-client
PyYAML
redis